The projectiles are now sensors and the Robot Projectile Sensor is removed
Before this update, each robot had its' own projectile sensor which was a n invisible sensor collider that followed the robot's position and rotation and was used so that the projectile colliders don't hit the robot hull collider directly so they don't push them.

But now I set the projectiles to be sensors instead and removed the robot projectile sensor.